+++ src/main/scala/characters/BlackMage.scala
@@ -1,5 +1,5 @@
 package characters
 
 /** Class for Magical Character: Black Mage */
-class BlackMage(name: String, healthPoints: Double, defense: Double, weight: Double, manaPoints: Double) extends AbstractMagicalCharacter(name, healthPoints, defense, weight, manaPoints) {
+class BlackMage(name: String, healthPoints: Int, defense: Int, weight: Int, manaPoints: Int) extends AbstractMagicalCharacter(name, healthPoints, defense, weight, manaPoints) {
 }
