+++ src/main/scala/characters/Ninja.scala
@@ -1,5 +1,5 @@
 package characters
 
 /** Class for Ninja Common Character */
-class Ninja(name: String, healthPoints: Double, defense: Double, weight: Double) extends CommonCharacter(name, healthPoints, defense, weight) {
+class Ninja(name: String, healthPoints: Int, defense: Int, weight: Int) extends CommonCharacter(name, healthPoints, defense, weight) {
 }
