+++ src/main/scala/characters/Paladin.scala
@@ -1,5 +1,5 @@
 package characters
 
 /** Class for Common Character: Paladin */
-class Paladin(name: String, healthPoints: Double, defense: Double, weight: Double) extends CommonCharacter(name, healthPoints, defense, weight) {
+class Paladin(name: String, healthPoints: Int, defense: Int, weight: Int) extends CommonCharacter(name, healthPoints, defense, weight) {
 }
