+++ src/main/scala/characters/Warrior.scala
@@ -1,5 +1,5 @@
 package characters
 
 /** Class for Common Character: Warrior */
-class Warrior(name: String, healthPoints: Double, defense: Double, weight: Double) extends CommonCharacter(name, healthPoints, defense, weight) {
+class Warrior(name: String, healthPoints: Int, defense: Int, weight: Int) extends CommonCharacter(name, healthPoints, defense, weight) {
 }
