+++ src/main/scala/characters/WhiteMage.scala
@@ -1,5 +1,5 @@
 package characters
 
 /** Class for Magical Character: White Mage */
-class WhiteMage(name: String, healthPoints: Double, defense: Double, weight: Double, manaPoints: Double) extends AbstractMagicalCharacter(name, healthPoints, defense, weight, manaPoints) {
+class WhiteMage(name: String, healthPoints: Int, defense: Int, weight: Int, manaPoints: Int) extends AbstractMagicalCharacter(name, healthPoints, defense, weight, manaPoints) {
 }
